# smartav_api/static/openapi.json
{
    "openapi": "3.0.3",
    "info": {
      "title": "SmartAV Swagger Document - OpenAPI 3.0",
      "version": "v1.0"
    },
...
    "paths": {
      "/face-recognition/config-database": {
        "post": {
          "tags": [
            "Face Recognition"
          ],
          "description": "Set the parameters for database connection",
          "operationId": "configDatabase",
          "requestBody": {
            "description": "Set the parameters for database connection",
            "content": {
              "application/json": {
                "schema": {
                  "required": [
                    "host",
                    "port",
                    "db_name",
                    "username",
                    "password"
                  ],
                  "type": "object",
                  "properties": {
                    "host": {
                      "type": "string",
                      "example": "<host domain/ip>"
                    },
                    "port": {
                      "type": "integer",
                      "example": 5000
                    },
                    "db_name": {
                      "type": "string",
                      "example": "face_recognition"
                    },
                    "username": {
                      "type": "string",
                      "example": "db_username"
                    },
                    "password": {
                      "type": "string",
                      "example": "db_password"
                    }
                  }
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "Successful operation",
              "content": {
                "text/plain": {
                  "schema": {
                    "type": "string",
                    "example": "Database configuration is done."
                  }
                }
              }
            },
            "400": {
              "description": "Invalid request",
              "content": {
                "text/plain": {
                  "schema": {
                    "type": "string",
                    "example": "\"host\" is missing"
                  }
                }
              }
            },
            "500": {
              "description": "Internal Server Error",
              "content": {
                "text/plain": {
                  "schema": {
                    "type": "string",
                    "example": "Database configuration is failed. Failed to create tables for face recognition module."
                  }
                }
              }
            }
          }
        }
      },
      "/face-recognition/clear-samples": {
        "post": {
          "tags": [
            "Face Recognition"
          ],
          "description": "Clear all feature vectors of the persons saved in database",
          "operationId": "clearSample",
          "responses": {
            "200": {
              "description": "All feature vectors have been removed from database successfully.",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "success": {
                        "type": "boolean",
                        "example": true
                      },
                      "msg": {
                        "type": "string",
                        "example": "Samples have been removed successfully"
                      }
                    }
                  }
                }
              }
            },
            "500": {
              "description": "Failed to clear the feature vectors",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "success": {
                        "type": "boolean",
                        "example": false
                      },
                      "msg": {
                        "type": "string",
                        "example": "Failed to clear the existing samples"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/face-recognition/update-samples": {
        "post": {
          "tags": [
            "Face Recognition"
          ],
          "description": "Add feature vectors of faces to database. <br><br> When you send an image for existing \"id\", the image will be appended to the \"id\" face. <br> In this way, each face can have multiple feature vectors, and all vectors will be used when we recognize the face.<br> This will increase the accuracy of our recognition module.",
          "operationId": "updateSamples",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "example": "person1"
                      },
                      "name": {
                        "type": "string",
                        "example": "person1"
                      },
                      "image": {
                        "type": "string",
                        "example": "data:image/jpeg;base64,/9j/4AAQSkZJRgABAQEAYABgAAD...."
                      },
                      "metadata": {
                        "type": "string",
                        "example": "mydomain.com/myobject1"
                      },
                      "action": {
                        "type": "string",
                        "example": "embedlink"
                      }
                    }
                  }
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "\\\"success\\\" will contain the \\\"id\\\"s that are saved successfully, while \\\"fail\\\" will contain \\\"id\\\"s that are failed.",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "success": {
                        "type": "array",
                        "items": {
                          "type": "string",
                          "example": "person1"
                        }
                      },
                      "fail": {
                        "type": "array",
                        "items": {
                          "type": "string",
                          "example": "person2"
                        }
                      }
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Invalid request",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "error": {
                        "type": "string",
                        "example": "Invalid request"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/face-recognition/update-metadata/{sample_id}": {
        "post": {
          "tags": [
            "Face Recognition"
          ],
          "description": "Update the metadata of the specific person",
          "operationId": "updateMetadata",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "required": [
                    "name",
                    "metadata"
                  ],
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string",
                      "example": "person1"
                    },
                    "metadata": {
                      "type": "string",
                      "example": "new metadata"
                    }
                  }
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "Successfully update the metadata for specific person.",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "success": {
                        "type": "string",
                        "example": "\"person1\" has been configured successfully."
                      }
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Invalid request. You might be missing one or more required fields.",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "error": {
                        "type": "string",
                        "example": "\"name\" is missing in the request"
                      }
                    }
                  }
                }
              }
            },
            "500": {
              "description": "Internal server error",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "error": {
                        "type": "string",
                        "example": "DB connection error occurred"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/face-recognition": {
        "post": {
          "tags": [
            "Face Recognition"
          ],
          "description": "Recognize the faces in the image.<br><br> \"min_distance\" is an optional field. The value of this field is related with the feature extraction model. <br> For example, it will be in [0, 1]. <br> Since we are using \"facenet\" model as default, the ideal threshold is 0.35.",
          "operationId": "faceRecognition",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "required": [
                    "image"
                  ],
                  "type": "object",
                  "properties": {
                    "image": {
                      "type": "string",
                      "example": "data:image/jpeg;base64,/9j/4AAQSkZJRgABAQEAYABgAAD...."
                    },
                    "min_distance": {
                      "type": "number",
                      "format": "float",
                      "example": 0.35
                    }
                  }
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "Returns the candidates most similar to the face in the input image from samples in the database.",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "example": "person1"
                      },
                      "name": {
                        "type": "string",
                        "example": "person1"
                      },
                      "metadata": {
                        "type": "string",
                        "example": "embedlink"
                      },
                      "bbox": {
                        "type": "string",
                        "example": "x1, y1, w, h"
                      }
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Invalid request",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "error": {
                        "type": "string",
                        "example": "Invalid request"
                      }
                    }
                  }
                }
              }
            },
            "500": {
              "description": "Internal server error",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "error": {
                        "type": "string",
                        "example": "Feature extraction node is not running."
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/set-story-gen-api-host": {
        "post": {
          "tags": [
            "Story Generator"
          ],
          "description": "You should set the host url of the story generator node.",
          "operationId": "setStoryGenAPIHost",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "required": [
                    "hostname"
                  ],
                  "type": "object",
                  "properties": {
                    "hostname": {
                      "type": "string",
                      "example": "http://localhost:8000"
                    }
                  }
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "Successfully set the host url",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "result": {
                        "type": "string",
                        "example": "success"
                      }
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Invalid request.",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "error": {
                        "type": "string",
                        "example": "Invalid request."
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/init-story-generator": {
        "post": {
          "tags": [
            "Story Generator"
          ],
          "description": "You should initialize the story generator before you call the other endpoints related to story generator.",
          "operationId": "initStoryGenerator",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "required": [
                    "prompt"
                  ],
                  "type": "object",
                  "properties": {
                    "prompt": {
                      "type": "string",
                      "example": "Your initlai text"
                    }
                  }
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "The story generator has been successfully initiated with the input prompt.",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "result": {
                        "type": "string",
                        "example": "generated text"
                      }
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Invalid request.",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "error": {
                        "type": "string",
                        "example": "Invalid request. The request should contain \"prompt\"."
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/set-story-gen-interval": {
        "post": {
          "tags": [
            "Story Generator"
          ],
          "description": "It will generate the story every time when the <code>/image-captioning</code> endpoint has been called 3 times.",
          "operationId": "setStoryGenInterval",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "interval"
                  ],
                  "properties": {
                    "interval": {
                      "type": "integer",
                      "example": 3
                    }
                  }
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "The interval has been successfully set.",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "result": {
                        "type": "string",
                        "example": "success"
                      }
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Invalid request. You might be missing \"interval\" in request.",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "error": {
                        "type": "string",
                        "example": "Invalid request."
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
...
    },
...
  }